requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
Django==6.0.3
djangorestframework==3.16.1
msgpack==1.1.2
pillow==12.1.1
redis==7.3.0
sqlparse==0.5.5
tzdata==2025.3
